Edit app to start
